+++ platform/src/common/components/Map/functions/useMapControls.jsx
